+++ packages/tendermint-light-client/update-client/Cargo.toml
@@ -16,9 +16,11 @@ ibc-core-commitment-types        = { workspace = true }
 ibc-primitives                   = { workspace = true }
 sha2                             = { workspace = true }
 thiserror                        = { workspace = true }
-serde                            = { workspace = true, features = ["derive"] }
-hex                              = { workspace = true }
-serde_json                       = { workspace = true, features = ["std"] }
 tendermint                       = { workspace = true }
 prost                            = { workspace = true }
+
+[dev-dependencies]
+serde                            = { workspace = true, features = ["derive"] }
+serde_json                       = { workspace = true, features = ["std"] }
+hex                              = { workspace = true }
