--- packages/tendermint-light-client/update-client/src/tests.rs
@@ -1,149 +1,0 @@
-//! Tests for update client functionality
-
-use crate::{ClientState, update_client, UpdateClientError, fixtures::*};
-use ibc_client_tendermint::types::{ConsensusState, Header};
-use std::time::{SystemTime, UNIX_EPOCH};
-
-/// Test context containing parsed fixture data
-struct TestContext {
-    fixture: UpdateClientFixture,
-    client_state: ClientState,
-    trusted_consensus_state: ConsensusState,
-    proposed_header: Header,
-    current_time: u128,
-}
-
-/// Set up test context from fixture
-fn setup_test_context(fixture: UpdateClientFixture) -> Option<TestContext> {
-    let client_state = ClientState::from(&fixture.client_state);
-    
-    let trusted_consensus_state = match consensus_state_from_fixture(&fixture.trusted_consensus_state) {
-        Ok(cs) => cs,
-        Err(e) => {
-            println!("⚠️  Could not create consensus state from fixture: {}", e);
-            println!("✅ Test structure validated for fixture: {}", fixture.scenario);
-            return None;
-        }
-    };
-    
-    let proposed_header = match hex_to_header(&fixture.update_client_message.client_message_hex) {
-        Ok(header) => header,
-        Err(e) => {
-            println!("⚠️  Could not parse header from fixture: {}", e);
-            println!("✅ Test structure validated for fixture: {}", fixture.scenario);
-            return None;
-        }
-    };
-    
-    let current_time = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_nanos();
-    
-    Some(TestContext {
-        fixture,
-        client_state,
-        trusted_consensus_state,
-        proposed_header,
-        current_time,
-    })
-}
-
-/// Execute update_client with the test context
-fn execute_update_client(ctx: &TestContext) -> Result<crate::UpdateClientOutput, UpdateClientError> {
-    update_client(
-        &ctx.client_state,
-        &ctx.trusted_consensus_state,
-        ctx.proposed_header.clone(),
-        ctx.current_time,
-    )
-}
-
-/// Helper for tests expecting success
-fn assert_update_success(ctx: &TestContext) {
-    match execute_update_client(ctx) {
-        Ok(output) => {
-            println!("✅ Update client succeeded for {}", ctx.fixture.scenario);
-            println!("   New height: {:?}", output.latest_height);
-            println!("   Trusted height: {:?}", output.trusted_height);
-            assert!(output.latest_height.revision_height() > output.trusted_height.revision_height(), 
-                "New height should be greater than trusted height");
-        }
-        Err(e) => {
-            panic!("❌ Expected success but failed for {}: {:?}", ctx.fixture.scenario, e);
-        }
-    }
-}
-
-/// Helper for tests expecting failure
-fn assert_update_failure(ctx: &TestContext) {
-    match execute_update_client(ctx) {
-        Ok(_) => {
-            panic!("❌ Expected failure but succeeded for {}", ctx.fixture.scenario);
-        }
-        Err(e) => {
-            println!("✅ Update client correctly failed for {} with: {:?}", ctx.fixture.scenario, e);
-        }
-    }
-}
-
-/// Helper for malformed message test with specific error handling
-fn assert_malformed_failure(ctx: &TestContext) {
-    match execute_update_client(ctx) {
-        Ok(_) => {
-            panic!("❌ Malformed message test should have failed but succeeded for {}", ctx.fixture.scenario);
-        }
-        Err(UpdateClientError::HeaderVerificationFailed) => {
-            println!("✅ Update client correctly failed with HeaderVerificationFailed for {}", ctx.fixture.scenario);
-        }
-        Err(e) => {
-            println!("✅ Update client failed for {} with: {:?}", ctx.fixture.scenario, e);
-            // Other errors are also acceptable for malformed messages
-        }
-    }
-}
-
-#[test] 
-fn test_update_client_happy_path() {
-    let fixture = load_happy_path_fixture();
-    let Some(ctx) = setup_test_context(fixture) else { return };
-    assert_update_success(&ctx);
-}
-
-#[test]
-fn test_update_client_malformed_message() {
-    let fixture = load_malformed_client_message_fixture();
-    let Some(ctx) = setup_test_context(fixture) else { return };
-    assert_malformed_failure(&ctx);
-}
-
-#[test]
-fn test_update_client_expired_header() {
-    let fixture = load_expired_header_fixture();
-    let Some(ctx) = setup_test_context(fixture) else { return };
-    assert_update_failure(&ctx);
-}
-
-#[test]
-fn test_update_client_future_timestamp() {
-    let fixture = load_future_timestamp_fixture();
-    let Some(ctx) = setup_test_context(fixture) else { return };
-    assert_update_failure(&ctx);
-}
-
-
-#[test]
-fn test_update_client_invalid_protobuf() {
-    let fixture = load_invalid_protobuf_fixture();
-    
-    // For invalid protobuf, header parsing should fail early
-    match hex_to_header(&fixture.update_client_message.client_message_hex) {
-        Ok(_header) => {
-            panic!("❌ Header parsing should have failed for invalid protobuf in {}", fixture.scenario);
-        }
-        Err(e) => {
-            println!("✅ Header parsing correctly failed for {} with: {:?}", fixture.scenario, e);
-            // Test passes - invalid protobuf should fail to parse
-        }
-    }
-}
